Module turbopack_cli_utils::issue

Structs

ConsoleUi

Logs emitted issues to console logs, deduplicating issues between peeks of the collected issues. The ConsoleUi can be shared and capture issues from multiple sources, with deduplication operating across all issues.

ConsoleUiVc

Vc for ConsoleUi

DisplayIssueState

DisplayIssueStateVc

Vc for DisplayIssueState

IssueSeverityCliOption

LogOptions

LogOptionsVc

Vc for LogOptions

Functions

format_issue

Type Definitions

ConsoleUiReadRef

see turbo_tasks::ReadRef

DisplayIssueStateReadRef

see turbo_tasks::ReadRef

GroupedIssues

LogOptionsReadRef

see turbo_tasks::ReadRef